refactor TaskArchiver avoid race condition
